+++ packages/io/x3d-serializer/src/index.js
@@ -22,11 +22,15 @@ import { stringify } from '@jscad/io-utils'
 const { geom2, geom3, path2, poly2, poly3 } = geometries
 
 const mimeType = 'model/x3d+xml'
+const defNames = new Map()
 
 /**
  * Serialize the give objects to X3D elements (XML).
  * @param {Object} options - options for serialization, REQUIRED
  * @param {Array} [options.color=[0,0,1,1]] - default color for objects
+ * @param {Number} [options.shininess=8/256] - x3d shininess for specular highlights
+ * @param {Boolean} [options.smooth=false] - use averaged vertex normals
+ * @param {Number} [options.decimals=1000] - multiplier before rounding to limit precision
  * @param {Boolean} [options.metadata=true] - add metadata to 3MF contents, such at CreationDate
  * @param {String} [options.unit='millimeter'] - unit of design; millimeter, inch, feet, meter or micrometer
  * @param {Function} [options.statusCallback] - call back function for progress ({ progress: 0-100 })
@@ -40,6 +44,8 @@ const mimeType = 'model/x3d+xml'
 const serialize = (options, ...objects) => {
   const defaults = {
     color: [0, 0, 1, 1.0], // default colorRGBA specification
+    shininess: 8 / 256,
+    smooth: false,
     decimals: 1000,
     metadata: true,
     unit: 'millimeter', // millimeter, inch, feet, meter or micrometer
@@ -59,9 +65,9 @@ const serialize = (options, ...objects) => {
   let body = ['X3D',
     {
       profile: 'Interchange',
-      version: '4.0',
+      version: '3.3',
       'xmlns:xsd': 'http://www.w3.org/2001/XMLSchema-instance',
-      'xsd:noNamespaceSchemaLocation': 'http://www.web3d.org/specifications/x3d-4.0.xsd'
+      'xsd:noNamespaceSchemaLocation': 'http://www.web3d.org/specifications/x3d-3.3.xsd'
     }
   ]
   if (options.metadata) {
@@ -87,7 +93,6 @@ ${stringify(body, 2)}`
 }
 
 const convertObjects = (objects, options) => {
-  let scene = ['Scene', {}]
   const shapes = []
   objects.forEach((object, i) => {
     options.statusCallback && options.statusCallback({ progress: 100 * i / objects.length })
@@ -107,7 +112,8 @@ const convertObjects = (objects, options) => {
       shapes.push(convertPath2(object, options))
     }
   })
-  scene = scene.concat(shapes)
+  const transform = ['Transform', { rotation: '1 0 0 -1.5708' }, ...shapes]
+  const scene = ['Scene', {}, transform]
   return [scene]
 }
 
@@ -117,9 +123,9 @@ const convertObjects = (objects, options) => {
 const convertPath2 = (object, options) => {
   const points = path2.toPoints(object).slice()
   if (points.length > 1 && object.isClosed) points.push(points[0])
-  const shape = ['Shape', {}, convertPolyline2D(poly2.create(points), options)]
+  const shape = ['Shape', shapeAttributes(object), convertPolyline2D(poly2.create(points), options)]
   if (object.color) {
-    shape.push(convertAppearance(object, options))
+    shape.push(convertAppearance(object, 'emissiveColor', options))
   }
   return shape
 }
@@ -132,13 +138,31 @@ const convertGeom2 = (object, options) => {
   const group = ['Group', {}]
   outlines.forEach((outline) => {
     if (outline.length > 1) outline.push(outline[0]) // close the outline for conversion
-    const shape = ['Shape', {}, convertPolyline2D(poly2.create(outline), options)]
+    const shape = ['Shape', shapeAttributes(object), convertPolyline2D(poly2.create(outline), options)]
     if (object.color) {
-      shape.push(convertAppearance(object, options))
+      shape.push(convertAppearance(object, 'emissiveColor', options))
     }
     group.push(shape)
   })
   return group
+}
+
+/*
+ * generate attributes for Shape node
+ */
+
+const shapeAttributes = (object, attributes = {}) => {
+  if (object.id) {
+    Object.assign(attributes, { DEF: checkDefName(object.id) })
+  }
+  return attributes
+}
+
+const checkDefName = (defName) => {
+  const count = defNames.get(defName) || 0
+  defNames.set(defName, count + 1)
+  if (count > 0) console.warn(`Warning: object.id set as DEF but not unique. ${defName} set ${count + 1} times.`)
+  return defName
 }
 
 /*
@@ -149,20 +173,32 @@ const convertPolyline2D = (object, options) => {
   return ['Polyline2D', { lineSegments }]
 }
 
-const convertAppearance = (object, options) => {
-  const diffuseColor = object.color.join(' ')
-  const emissiveColor = object.color.join(' ')
-  return ['Appearance', ['Material', { diffuseColor, emissiveColor }]]
+/*
+ * Convert color to Appearance
+ */
+const convertAppearance = (object, colorField, options) => {
+  const colorRGB = object.color.slice(0, 3)
+  const color = colorRGB.join(' ')
+  const transparency = roundToDecimals(1.0 - object.color[3], options)
+  const materialFields = { [colorField]: color, transparency }
+  if (colorField === 'diffuseColor') {
+    Object.assign(
+      materialFields,
+      { specularColor: '0.2 0.2 0.2', shininess: options.shininess })
+  }
+  return ['Appearance', ['Material', materialFields]]
 }
 
 /*
  * Convert the given object (geom3) to X3D source
  */
 const convertGeom3 = (object, options) => {
-  const shape = ['Shape', {}, convertMesh(object, options)]
+  const shape = ['Shape', shapeAttributes(object), convertMesh(object, options)]
+  let appearance = ['Appearance', {}, ['Material']]
   if (object.color) {
-    shape.push(convertAppearance(object, options))
+    appearance = convertAppearance(object, 'diffuseColor', options)
   }
+  shape.push(appearance)
   return shape
 }
 
@@ -176,7 +212,7 @@ const convertMesh = (object, options) => {
 
   const faceset = [
     'IndexedTriangleSet',
-    { ccw: 'true', colorPerVertex: 'false', solid: 'false', index: indexList },
+    { ccw: 'true', colorPerVertex: 'false', normalPerVertex: options.smooth, solid: 'false', index: indexList },
     ['Coordinate', { point: pointList }]
   ]
   if (!object.color) {
@@ -215,6 +251,8 @@ const convertToColor = (polygon, options) => {
   return `${color[0]} ${color[1]} ${color[2]}`
 }
 
+const roundToDecimals = (float, options) => Math.round(float * options.decimals) / options.decimals
+
 /*
  * This function converts the given polygons into three lists
  * - indexList : index of each vertex in the triangle (tuples)
@@ -236,9 +274,9 @@ const polygons2coordinates = (polygons, options) => {
 
       // add the vertex to the list of points (and index) if not found
       if (!vertexTagToCoordIndexMap.has(id)) {
-        const x = Math.round(vertex[0] * options.decimals) / options.decimals
-        const y = Math.round(vertex[1] * options.decimals) / options.decimals
-        const z = Math.round(vertex[2] * options.decimals) / options.decimals
+        const x = roundToDecimals(vertex[0], options)
+        const y = roundToDecimals(vertex[1], options)
+        const z = roundToDecimals(vertex[2], options)
         pointList.push(`${x} ${y} ${z}`)
         vertexTagToCoordIndexMap.set(id, pointList.length - 1)
       }