+++ packages/modeling/src/curves/bezier/length.d.ts
@@ -1,0 +1,3 @@
+import Bezier from './type'
+
+export function length(segments: number, bezier: Bezier): number