+++ packages/modeling/src/geometries/slice/create.d.ts
@@ -1,3 +1,4 @@
+import { Vec3 } from '../../maths/types'
 import { Slice } from './type'
 
 export function create(contours?: Array<Array<Vec3>>): Slice