+++ packages/modeling/src/maths/vec2/angle.d.ts
@@ -1,1 +1,1 @@
-import angleRadians from './angleRadians'
+export { angleRadians as angle } from './angleRadians'