+++ packages/modeling/src/operations/modifiers/index.js
@@ -7,3 +7,4 @@
  */
 export { generalize } from './generalize.js'
 export { snap } from './snap.js'
+export { retessellate } from './retessellate.js'