+++ packages/modeling/src/operations/modifiers/retessellate.d.ts
@@ -1,0 +1,3 @@
+import { Geom3 } from '../../geometries/types'
+
+export function retessellate(geometry: Geom3): Geom3