+++ packages/core/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.7](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/core@2.6.6...@jscad/core@2.6.7) (2023-06-27)
+
+**Note:** Version bump only for package @jscad/core
+
+
+
+
+
+## [2.6.6](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/core@2.6.5...@jscad/core@2.6.6) (2023-04-30)
+
+**Note:** Version bump only for package @jscad/core
+
+
+
+
+
+## [2.6.5](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/core@2.6.4...@jscad/core@2.6.5) (2022-11-26)
+
+**Note:** Version bump only for package @jscad/core
+
+
+
+
+
 ## [2.6.4](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/core@2.6.3...@jscad/core@2.6.4) (2022-09-23)
 
 **Note:** Version bump only for package @jscad/core