+++ packages/io/stl-deserializer/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.22](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/stl-deserializer@2.1.21...@jscad/stl-deserializer@2.1.22) (2023-06-27)
+
+**Note:** Version bump only for package @jscad/stl-deserializer
+
+
+
+
+
+## [2.1.21](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/stl-deserializer@2.1.20...@jscad/stl-deserializer@2.1.21) (2023-04-30)
+
+**Note:** Version bump only for package @jscad/stl-deserializer
+
+
+
+
+
+## [2.1.20](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/stl-deserializer@2.1.19...@jscad/stl-deserializer@2.1.20) (2022-11-26)
+
+**Note:** Version bump only for package @jscad/stl-deserializer
+
+
+
+
+
 ## [2.1.19](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/stl-deserializer@2.1.18...@jscad/stl-deserializer@2.1.19) (2022-08-21)
 
 **Note:** Version bump only for package @jscad/stl-deserializer