+++ packages/io/x3d-deserializer/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.5](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/x3d-deserializer@2.2.4...@jscad/x3d-deserializer@2.2.5) (2023-06-27)
+
+**Note:** Version bump only for package @jscad/x3d-deserializer
+
+
+
+
+
+## [2.2.4](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/x3d-deserializer@2.2.3...@jscad/x3d-deserializer@2.2.4) (2023-04-30)
+
+**Note:** Version bump only for package @jscad/x3d-deserializer
+
+
+
+
+
+## [2.2.3](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/x3d-deserializer@2.2.2...@jscad/x3d-deserializer@2.2.3) (2022-11-26)
+
+**Note:** Version bump only for package @jscad/x3d-deserializer
+
+
+
+
+
 ## [2.2.2](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/x3d-deserializer@2.2.1...@jscad/x3d-deserializer@2.2.2) (2022-08-21)
 
 **Note:** Version bump only for package @jscad/x3d-deserializer