+++ packages/web/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.5](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/web@2.6.4...@jscad/web@2.6.5) (2023-06-27)
+
+**Note:** Version bump only for package @jscad/web
+
+
+
+
+
+## [2.6.4](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/web@2.6.3...@jscad/web@2.6.4) (2023-04-30)
+
+**Note:** Version bump only for package @jscad/web
+
+
+
+
+
+## [2.6.3](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/web@2.6.2...@jscad/web@2.6.3) (2023-01-09)
+
+
+### Bug Fixes
+
+* **web:** add render hotkey message in editor ([1e99fa5](https://github.com/jscad/OpenJSCAD.org/commit/1e99fa5222dfd12153f8318f3b112261088a2fe0))
+
+
+
+
+
+## [2.6.2](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/web@2.6.1...@jscad/web@2.6.2) (2022-11-26)
+
+**Note:** Version bump only for package @jscad/web
+
+
+
+
+
 ## [2.6.1](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/web@2.6.0...@jscad/web@2.6.1) (2022-09-23)
 
 **Note:** Version bump only for package @jscad/web